Refactor show_goals_assists_buttons function, move Game class in game.py file
